Upgrade controllers routes override issues by moving some of them
