+++ config/packages/_easy_admin/easy_admin_esterenmaps.yaml
@@ -6,7 +6,7 @@ easy_admin:
         Factions:
             label: admin.maps.factions
             class: EsterenMaps\Entity\Factions
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 fields:
                     - { property: id, label: 'admin.entities.common.id' }
@@ -23,7 +23,7 @@ easy_admin:
         Maps:
             label: admin.maps.maps
             class: EsterenMaps\Entity\Maps
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - edit
@@ -65,7 +65,7 @@ easy_admin:
         Markers:
             label: admin.maps.markers
             class: EsterenMaps\Entity\Markers
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -106,7 +106,7 @@ easy_admin:
         MarkersTypes:
             label: admin.maps.markers_types
             class: EsterenMaps\Entity\MarkersTypes
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -121,7 +121,7 @@ easy_admin:
         Routes:
             label: admin.maps.routes
             class: EsterenMaps\Entity\Routes
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -148,7 +148,7 @@ easy_admin:
         RoutesTypes:
             label: admin.maps.routes_types
             class: EsterenMaps\Entity\RoutesTypes
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -165,7 +165,7 @@ easy_admin:
         Zones:
             label: admin.maps.zones
             class: EsterenMaps\Entity\Zones
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -201,7 +201,7 @@ easy_admin:
         ZonesTypes:
             label: admin.maps.zones_types
             class: EsterenMaps\Entity\ZonesTypes
-            controller: EsterenMaps\Controller\Admin\BaseMapAdminController
+            controller: Admin\CustomController\BaseMapAdminController
             list:
                 actions:
                     - '-delete'
@@ -220,7 +220,7 @@ easy_admin:
         TransportTypes:
             label: admin.maps.transports_types
             class: EsterenMaps\Entity\TransportTypes
-            controller: EsterenMaps\Controller\Admin\TransportTypesController
+            controller: Admin\CustomController\TransportTypesController
             list:
                 actions:
                     - '-delete'
